+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Changed the display of schemas for the "Data" tab and "Query" page 
+- SetTransform event: replaced the list of columns and types with a table 
 
 ## [0.61.2] - 2026-01-30
 ### Added